# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Give Away Voucher to 1000 lucky customers",
        "version": "1.0"
    },
    "paths": {
        "/api/customer-check": {
            "post": {
                "tags": [
                    "customer check"
                ],
                "summary": "check if customer eligible for the event",
                "description": "Return unclaimed voucher code",
                "operationId": "checkCustomer",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Customer unique identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success booked voucher code or already have voucher code"
                    },
                    "403": {
                        "description": "All voucher has been claimed or fully booked"
                    },
                    "404": {
                        "description": "customer not found"
                    },
                    "422": {
                        "description": "Validation error (customer_id needed) "
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/validate-photo": {
            "post": {
                "tags": [
                    "validate photo"
                ],
                "summary": "Claim voucher by uploading product's photo",
                "description": "This API will return claimed voucher code with customer data if success",
                "operationId": "validatePhoto",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Customer unique identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "nullable": false,
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success claimed voucher code or already claimed voucher code"
                    },
                    "403": {
                        "description": "Customer dont have voucher to claim or Image recognizition API fail"
                    },
                    "404": {
                        "description": "customer not found"
                    },
                    "422": {
                        "description": "Validation error (customer_id needed) "
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}